Refactor and clean-up
